Updates for support in Jenkinsfile
Previously, trying to use this plugin when referencing the global tool inside of a Jenkinsfile failed miserably.

I've made a few updates to get it working.

Example:

pipeline {
    agent any
    tools {
        dependencyCheck 'DependencyCheck5'
    }
    stages {
        stage('Example') {
            steps {
                sh 'dependency-check.sh --version'
            }
        }
    }
}

There are two issues that have been fixed with this.

First, is the Symbol annotation on DescriptorImpl. I think the "-" was main culprit. I updated the string value to be "dependencyCheck" (rather than "dependency-check").

Second, was not setting the PATH environment variable appropriately to reference the tool.

I also made a change to how it was determined if Jenkins was on a Windows node or not.