Fix build status indication in blue ocean build step
Because the call "build.setResult()" does only affect the overall build
status and not the build step status, the status on the build step was
previously not displayed correctly. By throwing an exception in case the
Thresholds are exceeded, the build step status correctly shows the failure.

If the build step status is not correctly shown, it's sometimes hard to
find the reason for a build failure, specially if there are many build steps.