JENKINS-68342: Compare findings against last build
This commit adds two new tables to the "Dependency-Check Results" page for each build:
1. all findings that are new in this build.
2. findings that appear to have been fixed (present in the previous build, but not this build).

The implementation depends on comparing Finding instances from the previous build with those
from the current build. To do so, `equals()` and `hashCode()` implementations have been added
to classes that compose the `Finding` class, which is the majority of changes introduced in
this commit. A unit test has been added to verify this behavior.

Two new endpoints have been added that calculate the deltas at runtime. Each of these endpoints
are used by a corresponding new table that is added to the `ResultAction` Jelly view.